requirements.txt
aiogram==3.7.0
anthropic==0.28.0
feedparser==6.0.11
httpx==0.27.0
python-dotenv==1.0.1
